requirements.txt
alembic==1.18.4
amqp==5.3.1
annotated-types==0.7.0
anyio==4.12.1
apifairy==1.5.1
apispec==6.9.0
ariadne==1.0.0
bidict==0.23.1
billiard==4.2.4
blinker==1.8.2
celery==5.6.2
celery-redbeat==2.3.3
certifi==2026.2.25
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
distro==1.9.0
Flask==3.0.3
Flask-Admin==2.0.2
Flask-HTTPAuth==4.8.0
Flask-JWT-Extended==4.7.1
Flask-Login==0.6.3
Flask-Mail==0.10.0
flask-marshmallow==1.4.0
Flask-Migrate==4.1.0
Flask-SocketIO==5.6.1
Flask-SQLAlchemy==3.1.1
graphql-core==3.2.8
greenlet==3.1.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
importlib_metadata==8.5.0
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.13.0
kombu==5.6.2
Mako==1.3.10
MarkupSafe==2.1.5
marshmallow==4.2.2
marshmallow-sqlalchemy==1.4.2
openai==2.29.0
packaging==26.0
prompt_toolkit==3.0.52
pydantic==2.12.5
pydantic_core==2.41.5
PyJWT==2.12.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-engineio==4.13.1
python-socketio==5.16.1
redis==7.3.0
simple-websocket==1.1.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.44
starlette==0.52.1
tenacity==9.1.4
tqdm==4.67.3
typing_extensions==4.15.0
typing-inspection==0.4.2
tzdata==2025.3
tzlocal==5.3.1
uvicorn==0.33.0
vine==5.1.0
wcwidth==0.6.0
webargs==8.7.1
Werkzeug==3.0.6
wsproto==1.3.2
WTForms==3.2.1
zipp==3.20.2